Made a bunch of changes so it's way better now
Don't remember exactly what's different, but the range of shades is now accurate to the original image. Making both images black and white makes them look basically identical. The number of shades can now be controlled with an argument in the posterize function. Images can now be selected with cmd argument